# openapi/taxdb.yaml
# TaxDB API OpenAPI Specification
# This file will be generated by scripts/gen_openapi.py
# Run `make gen-openapi` to generate this file

openapi: 3.0.0
info:
  title: TaxDB API
  version: 0.1.0
  description: |
    # Tax Document Database API
    
    This API provides access to tax documents from multiple jurisdictions (Belgium, Spain, Germany).
    
    ## Features
    
    * Text search across documents
    * Vector similarity search using embeddings
    * Hybrid search combining text and vector similarity
    * Document metadata retrieval
    * Similar document discovery
    
    ## Authentication
    
    API requests are rate-limited based on client IP address.
    
    ## Jurisdictions
    
    The following jurisdictions are supported:
    
    * BE - Belgium
    * ES - Spain
    * DE - Germany
  contact:
    name: TaxDB Team
    email: taxdb-team@example.com
    url: https://example.com/contact/
  license:
    name: Internal Use Only
  termsOfService: https://example.com/terms/
paths:
  /healthz:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: |
        Returns a simple status response to verify the API is running.
        
        Rate limit: 60 requests per minute per client IP.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheck'
  /search:
    get:
      tags:
        - Search
      summary: Search documents by text
      description: |
        This endpoint performs a text-based search across document titles and summaries.
        Results can be filtered by jurisdiction and are paginated.
        
        In local mode, uses ILIKE for text search.
        In cloud mode, uses Azure AI Search.
        
        Rate limit: 30 requests per minute per client IP.
        
        Examples:
          - Search for tax documents: `/search?q=tax`
          - Search for Belgian VAT documents: `/search?q=VAT&jurisdiction=BE`
          - Get page 2 with 20 results: `/search?q=tax&page=2&page_size=20`
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 500
          description: Search query text
        - name: jurisdiction
          in: query
          required: false
          schema:
            type: string
            enum: [BE, ES, DE]
          description: Filter by jurisdiction
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Page size (max 100)
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by (e.g., issue_date, title)
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (asc or desc)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
  /search/vector:
    post:
      tags:
        - Search
      summary: Search documents by vector similarity
      description: |
        This endpoint performs a semantic search using vector embeddings.
        It converts the query text to a vector embedding and finds documents with similar meaning,
        regardless of exact keyword matches.
        
        Results include a similarity score between 0 and 1, where 1 is most similar.
        
        Rate limit: 20 requests per minute per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorQuery'
            example:
              text: "tax implications for cross-border transactions"
              jurisdiction: "BE"
              page: 1
              page_size: 10
              max_distance: 0.5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorSearchResult'
        '500':
          description: Error generating embedding
          
  /search/hybrid:
    post:
      tags:
        - Search
      summary: Perform hybrid search combining vector similarity and text search
      description: |
        This endpoint combines the power of semantic search (vector similarity) with
        traditional keyword search (text similarity) for more accurate results.
        
        The weights control the balance between semantic and keyword matching:
        - vector_weight: Weight for semantic similarity (0.0 to 1.0)
        - text_weight: Weight for keyword similarity (0.0 to 1.0)
        
        Rate limit: 20 requests per minute per client IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HybridQuery'
            example:
              text: "VAT exemptions for financial services"
              jurisdiction: "ES"
              page: 1
              page_size: 10
              vector_weight: 0.7
              text_weight: 0.3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VectorSearchResult'
        '500':
          description: Error generating embedding
          
  /doc/{id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      description: |
        Returns document metadata and a SAS URL for the blob.
        The SAS URL is valid for 15 minutes and provides read-only access to the document.
        
        Document IDs follow the format: `{jurisdiction}:{date}:{identifier}`
        Example: `BE:20250101:123`
        
        Rate limit: 60 requests per minute per client IP.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          
  /doc/{id}/similar:
    get:
      tags:
        - Documents
      summary: Get documents similar to a given document
      description: |
        Uses vector similarity to find documents with similar content to the specified document.
        Results include a similarity score between 0 and 1, where 1 is most similar.
        
        This is useful for finding related documents or exploring a topic further.
        
        Rate limit: 30 requests per minute per client IP.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
        - name: jurisdiction
          in: query
          required: false
          schema:
            type: string
            enum: [BE, ES, DE]
          description: Filter by jurisdiction
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Maximum number of similar documents to return
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentWithScoreDTO'
        '404':
          description: Document not found
          
  /jurisdictions/{jurisdiction}/documents:
    get:
      tags:
        - Jurisdictions
      summary: Get documents by jurisdiction
      description: |
        Returns a paginated list of documents for the specified jurisdiction.
        Documents are sorted by issue date by default (newest first).
        
        Valid jurisdiction codes:
        - BE: Belgium
        - ES: Spain
        - DE: Germany
        
        Rate limit: 30 requests per minute per client IP.
      parameters:
        - name: jurisdiction
          in: path
          required: true
          schema:
            type: string
            enum: [BE, ES, DE]
          description: Jurisdiction code
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Page size (max 100)
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (asc or desc)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
components:
  schemas:
    HealthCheck:
      type: object
      properties:
        status:
          type: string
          example: ok
          
    SearchResult:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentDTO'
        total:
          type: integer
          description: Total number of matching documents
          
    VectorSearchResult:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentWithScoreDTO'
        total:
          type: integer
          description: Total number of matching documents
          
    DocumentDTO:
      type: object
      properties:
        id:
          type: string
          example: BE:20250101:123
        jurisdiction:
          type: string
          enum: [BE, ES, DE]
        source_system:
          type: string
        document_type:
          type: string
        title:
          type: string
        summary:
          type: string
          nullable: true
        issue_date:
          type: string
          format: date
        effective_date:
          type: string
          format: date
          nullable: true
        language_orig:
          type: string
        blob_url:
          type: string
        created_at:
          type: string
          format: date-time
          
    DocumentWithScoreDTO:
      type: object
      properties:
        document:
          $ref: '#/components/schemas/DocumentDTO'
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Similarity score (0-1, where 1 is most similar)
          
    VectorQuery:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Query text to search for
          minLength: 1
          maxLength: 1000
        jurisdiction:
          type: string
          enum: [BE, ES, DE]
          description: Filter by jurisdiction
        page:
          type: integer
          minimum: 1
          default: 1
          description: Page number (1-based)
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Page size (max 100)
        max_distance:
          type: number
          format: float
          nullable: true
          description: Maximum distance threshold
          
    HybridQuery:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Query text to search for
          minLength: 1
          maxLength: 1000
        jurisdiction:
          type: string
          enum: [BE, ES, DE]
          description: Filter by jurisdiction
        page:
          type: integer
          minimum: 1
          default: 1
          description: Page number (1-based)
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Page size (max 100)
        vector_weight:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.7
          description: Weight for vector similarity (0.0 to 1.0)
        text_weight:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.3
          description: Weight for text similarity (0.0 to 1.0)
  
  tags:
    - name: Health
      description: Health check endpoints
    - name: Search
      description: Document search endpoints
    - name: Documents
      description: Document retrieval endpoints
    - name: Jurisdictions
      description: Jurisdiction-specific endpoints